+++ dash_pydantic_form/pydantic_form_scripts.js
@@ -270,4 +270,12 @@ dash_clientside.pydf = {
     }
     return dash_clientside.no_update
   },
+  displayErrors: (errors, ids) => {
+    if (!errors) return Array(ids.length).fill(dash_clientside.no_update)
+
+    return ids.map(id => {
+      const fullPath = id.parent ? `${id.parent}:${id.field}` : id.field
+      return errors[fullPath]
+    })
+  }
 }